Extract Immutable
